update source code

+++ resources/tag.py
@@ -8,7 +8,7 @@ from models import TagModel, StoreModel, ItemModel
 
 blp = Blueprint("Tags", __name__, description = "Operations on tags")
 
-@blp.route("/store/<string:store_id>/tag")
+@blp.route("/store/<int:store_id>/tag")
 class TagsInStore(MethodView):
     @blp.response(200, TagSchema(many = True))
     def get(self, store_id):
@@ -30,20 +30,35 @@ class TagsInStore(MethodView):
         except SQLAlchemyError as e:
             abort(500, message = str(e))
 
-@blp.route("/tag/<string:tag_id>")
+@blp.route("/tag/<int:tag_id>")
 class Tag(MethodView):
     @blp.response(200, TagSchema)
     def get(self, tag_id):
         tag = TagModel.query.get_or_404(tag_id)
         return tag
 
+    @blp.response(
+        202,
+        description = "Deletes a tag if no item is tagged with it.",
+        example={"message": "Tag deleted."}
+    )
+    @blp.alt_response(404, description="Tag not found.")
+    @blp.alt_response(
+        400,
+        description = "Returned if the tag is assigned to one or more items. In this case, the tag is not deleted."
+    )
     def delete(self, tag_id):
         tag = TagModel.query.get_or_404(tag_id)
-        db.session.delete(tag)
-        db.session.commit()
-        return {"message": "Tag deleted!"}, 200
+        if not tag.items:
+            db.session.delete(tag)
+            db.session.commit()
+            return {"message": "Tag deleted!"}, 200
+        abort(
+            400,
+            message = "Could not delete tag. Make sure tag is not associated with any items, then try again."
+        )
 
-@blp.route("/item/<string:item_id>/tag/<string:tag_id>")
+@blp.route("/item/<int:item_id>/tag/<int:tag_id>")
 class LinkTagsToItem(MethodView):
     def post(self, item_id, tag_id):
         item = ItemModel.query.get_or_404(item_id)
@@ -55,6 +70,19 @@ class LinkTagsToItem(MethodView):
             db.session.add(item)
             db.session.commit()
             return tag
+        except SQLAlchemyError:
+            abort(500, message="An error occurred while inserting the tag.")
+
+    def delete(self, item_id, tag_id):
+        item = ItemModel.query.get_or_404(item_id)
+        tag = TagModel.query.get_or_404(tag_id)
+        
+        item.tags.remove(tag)
+
+        try:
+            db.session.add(item)
+            db.session.commit()
+            return {"message": "Item removed from tag", "item": item, "tag": tag}
         except SQLAlchemyError:
             abort(500, message="An error occurred while inserting the tag.")
